Cleaned & refactored application
- Changed the names of the profile routes
- Changed the names of the profile files
- Renamed variables in different files
- Added line breaks and organized code in files
- Converted classless div wrappers into Fragments
