+++ src/components/Header/Header.js
@@ -5,8 +5,8 @@ import Navbar from 'react-bootstrap/Navbar'
 const authenticatedOptions = (
   <Fragment>
     <Nav.Link href="#chat">Chat</Nav.Link>
-    <Nav.Link href="#create-profile">Profiles</Nav.Link>
-    <Nav.Link href="#index-profile">See Your Profiles</Nav.Link>
+    <Nav.Link href="#profile/create">Create Profile</Nav.Link>
+    <Nav.Link href="#profiles">Profiles</Nav.Link>
     <Nav.Link href="#change-password">Change Password</Nav.Link>
     <Nav.Link href="#sign-out">Sign Out</Nav.Link>
   </Fragment>
