+++ src/components/Profile/EditProfile.js
@@ -1,10 +1,13 @@
-import React, { Component } from 'react'
+import React, { Component, Fragment } from 'react'
 import { Redirect } from 'react-router-dom'
-import { createProfile } from '../../api/profiles'
+
+import { updateProfile } from '../../api/profiles'
 import ProfileForm from './ProfileForm'
+
 class ProfileCreate extends Component {
   constructor (props) {
     super(props)
+
     // initially our profiles title and director will be empty until they are filled in
     this.state = {
       profile: {
@@ -12,42 +15,36 @@ class ProfileCreate extends Component {
         avatar: ''
       },
       // createdId will be null, until we successfully create a profile
-      created: false
+      updated: false
     }
   }
 
   handleSubmit = event => {
     event.preventDefault()
+
     const { user, msgAlert } = this.props
     const { profile } = this.state
+
     // create a profile, pass it the profile data and the user for its token
-    createProfile(profile, user)
+    updateProfile(profile, user)
       // set the createdId to the id of the profile we just created
       // .then(res => this.setState({ createdId: res.data.profile._id }))
       .then(res => {
-        this.setState({ created: true })
+        this.setState({ updated: true })
         // pass the response to the next .then so we can show the title
         return res
       })
       .then(res => msgAlert({
-        heading: 'Created Profile Successfully',
-        message: `Profile has been created successfully. Go Chat ${res.data.profile.nickname}!`,
+        heading: 'Profile Successfully Updated',
+        message: 'Profile has been successfully updated.',
         variant: 'success'
       }))
       .catch(error => {
-        if (error.response.status === 422) {
-          msgAlert({
-            heading: 'Failed to Create Profile',
-            message: 'A user can not have more than 1 profile',
-            variant: 'danger'
-          })
-        } else {
-          msgAlert({
-            heading: 'Failed to Create Profile',
-            message: 'Could not create profile with error: ' + error.message,
-            variant: 'danger'
-          })
-        }
+        msgAlert({
+          heading: 'Failed to Create Profile',
+          message: 'Could not create profile with error: ' + error.message,
+          variant: 'danger'
+        })
       })
   }
 
@@ -57,6 +54,7 @@ class ProfileCreate extends Component {
     // to ensure the properties are not set to null after handleChange is finished
     // we must call event.persist
     event.persist()
+
     this.setState(state => {
       // return our state changge
       return {
@@ -70,22 +68,24 @@ class ProfileCreate extends Component {
 
   render () {
     // destructure our profile and createdId state
-    const { profile, created } = this.state
+    const { profile, updated } = this.state
     // if the profile has been created and we set its id
-    if (created) {
+    if (updated) {
       // redirect to the profiles show page
       return <Redirect to={`/profiles/${profile.nickname}`} />
     }
+
     return (
-      <div>
-        <h3>Create Profile</h3>
+      <Fragment>
+        <h3>Edit Profile</h3>
         <ProfileForm
           profile={profile}
           handleChange={this.handleChange}
           handleSubmit={this.handleSubmit}
         />
-      </div>
+      </Fragment>
     )
   }
 }
+
 export default ProfileCreate
